+++ src/main/java/com/github/isa1412/detectordsbot/command/Command.java
@@ -1,7 +1,6 @@
 package com.github.isa1412.detectordsbot.command;
 
-import net.dv8tion.jda.api.events.GenericEvent;
-import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 
 /**
  * Command interface for handling discord bot commands.
@@ -11,7 +10,7 @@ public interface Command {
     /**
      * Main method, which is executing command logic.
      *
-     * @param event provided {@link GenericEvent} object with all the needed data for command.
+     * @param event provided {@link SlashCommandInteractionEvent} object with all the needed data for command.
      */
-    void execute(MessageReceivedEvent event);
+    void execute(SlashCommandInteractionEvent event);
 }