+++ src/main/java/com/github/isa1412/detectordsbot/command/CommandUtils.java
@@ -1,8 +1,10 @@
 package com.github.isa1412.detectordsbot.command;
 
 import com.github.isa1412.detectordsbot.repository.entity.id.MemberId;
-import net.dv8tion.jda.api.entities.MessageChannel;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
+
+import static java.util.Objects.requireNonNull;
 
 /**
  * Utils class for Commands.
@@ -10,33 +12,13 @@ import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 public class CommandUtils {
 
     /**
-     * Retrieve {@link MessageChannel} from {@link MessageReceivedEvent} object.
-     *
-     * @param event provided {@link MessageReceivedEvent}
-     * @return {@link MessageChannel} from the provided {@link MessageReceivedEvent} object.
-     */
-    public static MessageChannel getChannel(MessageReceivedEvent event) {
-        return event.getChannel();
-    }
-
-    /**
-     * Retrieve text of the message from {@link MessageReceivedEvent} object.
-     *
-     * @param event provided {@link MessageReceivedEvent}
-     * @return the text of the message from the provided {@link MessageReceivedEvent} object.
-     */
-    public static String getMessage(MessageReceivedEvent event) {
-        return event.getMessage().getContentDisplay();
-    }
-
-    /**
      * Retrieve users id of the message from {@link MessageReceivedEvent} object.
      *
      * @param event provided {@link MessageReceivedEvent}
      * @return the users id of the message from the provided {@link MessageReceivedEvent} object.
      */
-    public static String getUserId(MessageReceivedEvent event) {
-        return event.getAuthor().getId();
+    public static String getUserId(SlashCommandInteractionEvent event) {
+        return event.getUser().getId();
     }
 
     /**
@@ -45,8 +27,8 @@ public class CommandUtils {
      * @param event provided {@link MessageReceivedEvent}
      * @return the guilds id of the message from the provided {@link MessageReceivedEvent} object.
      */
-    public static String getGuildId(MessageReceivedEvent event) {
-        return event.getGuild().getId();
+    public static String getGuildId(SlashCommandInteractionEvent event) {
+        return requireNonNull(event.getGuild()).getId();
     }
 
     /**
@@ -55,7 +37,7 @@ public class CommandUtils {
      * @param event provided {@link MessageReceivedEvent}
      * @return the members id of the message from the provided {@link MessageReceivedEvent} object.
      */
-    public static MemberId getMemberId(MessageReceivedEvent event) {
+    public static MemberId getMemberId(SlashCommandInteractionEvent event) {
         return new MemberId(getUserId(event), getGuildId(event));
     }
 
@@ -65,8 +47,8 @@ public class CommandUtils {
      * @param event provided {@link MessageReceivedEvent}
      * @return the unix timestamp of the message from the provided {@link MessageReceivedEvent} object.
      */
-    public static long getTimestamp(MessageReceivedEvent event) {
+    public static long getTimestamp(SlashCommandInteractionEvent event) {
         long rechargeTime = 86400;
-        return event.getMessage().getTimeCreated().toInstant().getEpochSecond() + rechargeTime;
+        return event.getTimeCreated().toInstant().getEpochSecond() + rechargeTime;
     }
 }