+++ src/main/java/com/github/isa1412/detectordsbot/command/StopCommand.java
@@ -4,10 +4,8 @@ import com.github.isa1412.detectordsbot.repository.entity.id.MemberId;
 import com.github.isa1412.detectordsbot.service.MemberService;
 import com.github.isa1412.detectordsbot.service.ResponseGenerateService;
 import com.github.isa1412.detectordsbot.service.SendBotMessageService;
-import net.dv8tion.jda.api.entities.MessageChannel;
-import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 
-import static com.github.isa1412.detectordsbot.command.CommandUtils.getChannel;
 import static com.github.isa1412.detectordsbot.command.CommandUtils.getMemberId;
 
 /**
@@ -26,21 +24,20 @@ public class StopCommand implements Command {
     }
 
     @Override
-    public void execute(MessageReceivedEvent event) {
+    public void execute(SlashCommandInteractionEvent event) {
         MemberId memberId = getMemberId(event);
-        MessageChannel channel = getChannel(event);
 
         memberService.findById(memberId).ifPresentOrElse(
                 member -> {
                     if (member.isActive()) {
                         member.setActive(false);
                         memberService.save(member);
-                        messageService.sendMessage(channel, responseService.getOutGameResponse());
+                        messageService.sendReply(event, responseService.getOutGameResponse());
                     } else {
-                        messageService.sendMessage(channel, responseService.getAlreadyOutResponse());
+                        messageService.sendReply(event, responseService.getAlreadyOutResponse());
                     }
                 },
-                () -> messageService.sendMessage(channel, responseService.getNotMemberResponse())
+                () -> messageService.sendReply(event, responseService.getNotMemberResponse())
         );
     }
 }