+++ src/test/java/com/github/isa1412/detectordsbot/service/SendBotMessageServiceTest.java
@@ -1,7 +1,9 @@
 package com.github.isa1412.detectordsbot.service;
 
 import net.dv8tion.jda.api.entities.MessageChannel;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.requests.restaction.MessageAction;
+import net.dv8tion.jda.api.requests.restaction.interactions.ReplyCallbackAction;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -30,5 +32,20 @@ class SendBotMessageServiceTest {
 
         //then
         Mockito.verify(channel).sendMessage(message);
+    }
+
+    @Test
+    public void shouldProperlySendReply() {
+        //given
+        String message = "test_message";
+        SlashCommandInteractionEvent event = Mockito.mock(SlashCommandInteractionEvent.class);
+        ReplyCallbackAction action = Mockito.mock(ReplyCallbackAction.class);
+        Mockito.when(event.reply(message)).thenReturn(action);
+
+        //when
+        messageService.sendReply(event, message);
+
+        //then
+        Mockito.verify(event).reply(message);
     }
 }